+++ changelog.md
@@ -1,1 +1,1 @@
-- Added zoomable third person view
+- Farmer villagers can now harvest canola